fix(rtk-db,rtk-index): support rusqlite 0.40 FromSql/ToSql tightening (#65)
rusqlite 0.40 dropped the blanket FromSql/ToSql impl for usize.
Read row values as i64 and cast at the boundary instead of binding
usize directly, so the workspace can move past 0.32 (supersedes #59).

+++ rtk/rtk-db/src/tracking.rs
@@ -698,7 +698,7 @@ pub fn get_recent_logs(limit: usize) -> Result<Vec<CommandLog>> {
          FROM tracking ORDER BY id DESC LIMIT ?1"
     )?;
 
-    let rows = stmt.query_map(params![limit], |r| {
+    let rows = stmt.query_map(params![limit as i64], |r| {
         Ok(CommandLog {
             id: r.get(0)?,
             cmd: r.get(1)?,
